requirements.txt

fastapi==0.115.0
torch==2.4.1
torchvision==0.19.1
torchvision==0.19.1+cu121
torch-geometric==2.6.1
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-cluster==1.6.3
torch-spline-conv==1.2.2
numpy==2.1.2
scipy==1.14.1
scikit-learn==1.5.2
tifffile==2024.8.30
Pillow==10.4.0
opencv-python==4.10.0.84
tqdm==4.66.5
PyYAML==6.0.2
transformers==4.45.1